Devices subcomponent
