+++ react-native/react/tabs/devices/devices-render.mobile.js
